# Enabling linking in the ROCM/CUDA compiler targets.