--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPULinkExecutables.cpp
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPULinkExecutables.cpp
@@ -1,26 +1,62 @@
-// Copyright 2022 The IREE Authors
+// Copyright 2024 The IREE Authors
 //
 // Licensed under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Codegen/LLVMCPU/Passes.h"
+#include "iree/compiler/Codegen/LLVMGPU/Passes.h"
 #include "iree/compiler/Codegen/Utils/LinkingUtils.h"
 #include "iree/compiler/Utils/ModuleUtils.h"
 #include "llvm/Support/FormatVariadic.h"
+#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir::iree_compiler {
 
-#define GEN_PASS_DEF_LLVMCPULINKEXECUTABLESPASS
-#include "iree/compiler/Codegen/LLVMCPU/Passes.h.inc"
+#define GEN_PASS_DEF_LLVMGPULINKEXECUTABLESPASS
+#include "iree/compiler/Codegen/LLVMGPU/Passes.h.inc"
 
 namespace {
 
-struct LLVMCPULinkExecutablesPass
-    : public impl::LLVMCPULinkExecutablesPassBase<LLVMCPULinkExecutablesPass> {
-  using impl::LLVMCPULinkExecutablesPassBase<
-      LLVMCPULinkExecutablesPass>::LLVMCPULinkExecutablesPassBase;
+static bool isAddressSpacePrivate(uint32_t addressSpace) {
+  return addressSpace == /*local*/ 3 || addressSpace == /*private*/ 5;
+}
+
+static SymbolTable::Visibility
+convertLinkageToVisibility(LLVM::Linkage linkage) {
+  switch (linkage) {
+  case LLVM::Linkage::Private:
+    return SymbolTable::Visibility::Private;
+  case LLVM::Linkage::External:
+    return SymbolTable::Visibility::Public;
+  default:
+    return SymbolTable::Visibility::Public;
+  }
+}
+
+// Returns true if we are allowed to rename |op| as part of merging.
+// The LLVMGPU lowering is super careful about assigning linkage so we err on
+// the side of renaming (as 100% of usage today does not reference external
+// things).
+static bool allowRenamingPrivateLLVMSymbols(Operation *op) {
+  if (auto globalOp = dyn_cast<LLVM::GlobalOp>(op)) {
+    if (isAddressSpacePrivate(globalOp.getAddrSpace())) {
+      return true;
+    }
+    return convertLinkageToVisibility(globalOp.getLinkage()) ==
+           SymbolTable::Visibility::Private;
+  } else if (auto funcOp = dyn_cast<LLVM::LLVMFuncOp>(op)) {
+    return convertLinkageToVisibility(funcOp.getLinkage()) ==
+           SymbolTable::Visibility::Private;
+  }
+  return SymbolTable::getSymbolVisibility(op) ==
+         SymbolTable::Visibility::Private;
+}
+
+struct LLVMGPULinkExecutablesPass
+    : public impl::LLVMGPULinkExecutablesPassBase<LLVMGPULinkExecutablesPass> {
+  using impl::LLVMGPULinkExecutablesPassBase<
+      LLVMGPULinkExecutablesPass>::LLVMGPULinkExecutablesPassBase;
   void runOnOperation() override {
     auto moduleOp = getOperation();
     auto moduleBuilder = OpBuilder::atBlockBegin(moduleOp.getBody());
@@ -67,7 +103,8 @@ struct LLVMCPULinkExecutablesPass
       auto mergeModuleFn = [](mlir::ModuleOp sourceInnerModule,
                               mlir::ModuleOp linkedInnerModule,
                               DenseMap<StringRef, Operation *> &symbolMap) {
-        return mergeModuleInto(sourceInnerModule, linkedInnerModule, symbolMap);
+        return mergeModuleInto(sourceInnerModule, linkedInnerModule, symbolMap,
+                               allowRenamingPrivateLLVMSymbols);
       };
 
       // Try linking together all executables in moduleOp.
@@ -79,6 +116,5 @@ struct LLVMCPULinkExecutablesPass
     }
   }
 };
-
 } // namespace
 } // namespace mlir::iree_compiler