--- a/compiler/src/iree/compiler/Codegen/LLVMGPU/Passes.td
+++ b/compiler/src/iree/compiler/Codegen/LLVMGPU/Passes.td
@@ -66,6 +66,11 @@ def ExtractAddressComputationGPUPass: Pass<"extract-address-computation-gpu"> {
   ];
 }
 
+def LLVMGPUAssignConstantOrdinalsPass :
+    Pass<"iree-llvmgpu-assign-constant-ordinals", "IREE::HAL::ExecutableVariantOp"> {
+  let summary = "Assigns executable constant ordinals across all LLVMGPU variants.";
+}
+
 def LLVMGPUCastAddressSpaceFunctionPass :
     Pass<"iree-llvmgpu-cast-address-space-function", "ModuleOp"> {
   let summary = "Cast address space to generic in CallOp and FuncOp";
@@ -95,6 +100,18 @@ def LLVMGPUConvolutionToIGEMMPass :
     "iree_compiler::IREE::Codegen::IREECodegenDialect",
     "iree_compiler::IREE::GPU::IREEGPUDialect",
     "iree_compiler::IREE::LinalgExt::IREELinalgExtDialect"
+  ];
+}
+
+def LLVMGPULinkExecutablesPass :
+    Pass<"iree-llvmgpu-link-executables", "mlir::ModuleOp"> {
+  let summary = "Links LLVMGPU HAL executables within the top-level program module.";
+  let options = [
+    Option<
+      "target", "target",
+      "std::string", "",
+      "Target backend name whose executables will be linked by this pass."
+    >,
   ];
 }