@benvanik 2024-10-29
darn, I forgot the LLVM dialect doesn't implement symbol op visibility :(
`  <unknown>:0: error: multiple public symbols with the name: __shared_memory__`
I'll fix that.