@benvanik 2024-10-29
I was convinced this would be slower to compile but assuming our bot timing is in a consistent ballpark it's not much worse and may even be better in some cases. Or the machine is random, but it's at least not an order of magnitude off. VMFB sizes improve (a lot less ELF/metadata waste) and I'm sure startup time is significantly improved.

Without this change:
```
experimental/regression_suite/shark-test-suite-models/sdxl/test_unet.py::test_compile_punet_int8_fp16_rocm 
Compilation succeeded in 131.5626926422119s
experimental/regression_suite/shark-test-suite-models/sd3/test_vae.py::test_compile_vae_rocm 
Compilation succeeded in 122.2923686504364s
INFO     root:benchmark_sdxl_rocm.py:245 Scheduled Unet Binary Size: 1917726 bytes (golden binary size 2280000 bytes)
INFO     root:benchmark_sdxl_rocm.py:303 VAE Decode Binary Size: 793271 bytes (golden binary size 840000 bytes)
```

With this change:
```
experimental/regression_suite/shark-test-suite-models/sdxl/test_unet.py::test_compile_punet_int8_fp16_rocm 
Compilation succeeded in 93.87470889091492s
experimental/regression_suite/shark-test-suite-models/sd3/test_vae.py::test_compile_vae_rocm 
Compilation succeeded in 126.39901638031006s
INFO     root:benchmark_sdxl_rocm.py:245 Scheduled Unet Binary Size: 1319726 bytes (golden binary size 2280000 bytes)
INFO     root:benchmark_sdxl_rocm.py:303 VAE Decode Binary Size: 490445 bytes (golden binary size 840000 bytes)
```